+++ sst/elements/memHierarchy/trivialCPU.h
@@ -66,7 +66,7 @@ private:
 	int workPerCycle;
 	int commFreq;
 	bool do_write;
-	uint32_t maxAddr;
+	uint64_t maxAddr;
 	uint64_t num_reads_issued, num_reads_returned;
     uint64_t noncacheableRangeStart, noncacheableRangeEnd;
     uint64_t clock_ticks;
